Refactor "log" pkg to "logmetrics"
Adds ReportNow to logging
