+++ cmd/summarizer/BUILD.bazel
@@ -23,7 +23,7 @@ go_library(
         "//pkg/pubsub:go_default_library",
         "//pkg/summarizer:go_default_library",
         "//util/gcs:go_default_library",
-        "//util/metrics:go_default_library",
+        "//util/metrics/logmetrics:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go_pubsub//:go_default_library",
         "@org_golang_google_api//option:go_default_library",
