+++ util/metrics/BUILD.bazel
@@ -18,7 +18,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//util/metrics/log:all-srcs",
+        "//util/metrics/logmetrics:all-srcs",
         "//util/metrics/prometheus:all-srcs",
     ],
     tags = ["automanaged"],
