+++ util/metrics/logmetrics/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["log.go"],
-    importpath = "github.com/GoogleCloudPlatform/testgrid/util/metrics/log",
+    importpath = "github.com/GoogleCloudPlatform/testgrid/util/metrics/logmetrics",
     visibility = ["//visibility:public"],
     deps = [
         "//util/metrics:go_default_library",
